% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{filtered_lr_eg}
\alias{filtered_lr_eg}
\title{Example for filtered_lr}
\format{
An object of class \code{data.frame} with 5904 rows and 12 columns.
}
\usage{
filtered_lr_eg
}
\description{
Example for filtered_lr
}
\keyword{datasets}
