% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{matrix_object}
\alias{matrix_object}
\title{Example for matrix object}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 223 rows and 144 columns.
}
\usage{
matrix_object
}
\description{
Example for matrix object
}
\keyword{datasets}
