% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata_eg}
\alias{metadata_eg}
\title{Example for metadata}
\format{
An object of class \code{data.frame} with 163 rows and 8 columns.
}
\usage{
metadata_eg
}
\description{
Example for metadata
}
\keyword{datasets}
