% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_vis_sample.R
\name{pca_sample}
\alias{pca_sample}
\title{Generate PCA of LR Interaction Scores}
\usage{
pca_sample(
  lr_scores,
  metadata,
  score = "normalized",
  selected_sender = NULL,
  selected_receiver = NULL,
  color_by = NULL,
  n_components = 2
)
}
\arguments{
\item{lr_scores}{Data frame containing LR interaction scores per sample (data frame).}

\item{metadata}{Data frame containing sample metadata (data frame).}

\item{score}{Character string indicating which score to use: "normalized" (default) or "raw" .}

\item{selected_sender}{Specific sender cell type to filter, default is None (use all) (character).}

\item{selected_receiver}{Specific receiver cell type to filter, default is None (use all) (character).}

\item{color_by}{\code{metadata} column name to color points in PCA plot (character).}

\item{n_components}{Number of principal components to extract (numeric, default: 2).}
}
\value{
A list with two elements: the first is a ggplot2 PCA scatter plot and the second is the PCA results data frame.
}
\description{
This function performs principal component analysis (PCA) on ligand-receptor (LR) interaction scores across samples,
and generates a scatter plot of the first two principal components. Optionally, sample metadata can be used to color the points.
}
\examples{
# PCA of LR Interaction Scores
data(lr_scores_eg)
data(metadata_eg)
res <- pca_sample(lr_scores_eg, metadata_eg, score = "normalized",
  selected_sender = "Cardiac", selected_receiver = "Perivascular", color_by = "Age_group")
print(res$pca_plot)
}
