% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lr_scores_eg}
\alias{lr_scores_eg}
\title{Example for lr_scores}
\format{
An object of class \code{data.frame} with 377006 rows and 15 columns.
}
\usage{
lr_scores_eg
}
\description{
Example for lr_scores
}
\keyword{datasets}
